+++ cvxpy/expressions/constants/constant.py
@@ -33,7 +33,6 @@ class Constant(Leaf):
             self._value = intf.DEFAULT_SPARSE_INTERFACE.const_to_matrix(value)
             self._sparse = True
         else:
-            intf.warn_1D_array(value)
             self._value = intf.DEFAULT_INTERFACE.const_to_matrix(value)
             self._sparse = False
         # Set DCP attributes.