+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Version 2.13 (next release)
 ---------------------------
-* TODO get rid of NumPy 1D array warning.
+* Got rid of NumPy 1D array warning.
 * Made vstack and hstack not create new variables.
 * Replaced memoize with lazyprop in constraints.